Добавлено исключение, все double -> short

+++ src/CheeseBurger.java
@@ -1,11 +1,13 @@
+import BurgerFile.Burger;
+
 public class CheeseBurger implements Burger {
     public String getDescription(){
         return "Чизбургер";
     }
-    public double getCost(){
-        return 68.00;
+    public short getCost(){
+        return (short) 68.00;
     }
-    public double getCaloricContent(){
-        return 299.00;
+    public short getCaloricContent(){
+        return (short) 299.00;
     }
 }
