+++ src/DobryCola.java
@@ -1,12 +1,14 @@
+import DrinkFile.Drink;
+
 public class DobryCola implements Drink {
 
     public String getDescription(){
         return "Добрый кола";
     }
-    public double getCost(){
-        return 89.00;
+    public short getCost(){
+        return 89;
     }
-    public double getCaloricContent(){
-        return 106.00;
+    public short getCaloricContent(){
+        return 106;
     }
 }
